+++ environments/dibothrosuchus/envs/__init__.py
@@ -1,0 +1,3 @@
+from .dibothrosuchus_env import DibothrosuchusEnv
+
+__all__ = ["DibothrosuchusEnv"]